Move legacy test code out of the core module
