+++ libcaf_core/caf/all.hpp
@@ -70,7 +70,6 @@
 #include "caf/result.hpp"
 #include "caf/resumable.hpp"
 #include "caf/scheduler/abstract_coordinator.hpp"
-#include "caf/scheduler/test_coordinator.hpp"
 #include "caf/scoped_actor.hpp"
 #include "caf/scoped_execution_unit.hpp"
 #include "caf/sec.hpp"
